backend/requirements.txt
# FastAPI framework
fastapi==0.115.0

# ASGI server for running FastAPI
uvicorn[standard]==0.32.0

# WebSocket support
websockets==11.0.3


# Terminal text color and formatting
colorama==0.4.6

# LangChain libraries for AI/ML tasks
langchain-experimental==0.3.3
langchain-community==0.3.7
langchain-core==0.3.18
langchain-openai==0.2.8
langgraph==0.2.48

# Data validation and settings management
pydantic==2.9.2

# Environment variable management
python-dotenv==1.0.1

# GitHub API wrapper
PyGithub==2.1.1

# HTTP client for Python
requests==2.31.0

# JSON Web Token implementation
PyJWT==2.8.0

# YAML parser and emitter
PyYAML==6.0.1
